# api/http/openapi.yaml
openapi: "3.0.3"
info:
  title: Breakfix API
  version: "1.0.0"
  description: SRE/DevOps interview practice platform
servers:
  - url: http://localhost:9090/api
    description: Local dev
paths:
...
  /challenges/{id}/assistant:
    get:
      operationId: getChallengeAssistant
      summary: Read or create the assistant conversation for the active challenge environment
      tags: [assistant]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Environment-scoped assistant conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantConversation"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /challenges/{id}/assistant/messages:
    post:
      operationId: sendChallengeAssistantMessage
      summary: Send a message to the active challenge assistant and receive SSE events
      tags: [assistant]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantMessageRequest"
      responses:
        "200":
          description: Server-sent assistant status, tool, delta, complete, and error events
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AssistantEvidence:
      type: object
      required: [kind, label]
      properties:
        kind:
          type: string
        label:
          type: string
    AssistantMessage:
      type: object
      required: [id, role, content, created_at]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/AssistantEvidence"
        created_at:
          type: string
    AssistantConversation:
      type: object
      required: [id, challenge_id, messages]
      properties:
        id:
          type: string
        challenge_id:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AssistantMessage"
    AssistantTerminalContext:
      type: object
      required: [windows]
      properties:
        node:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,31}$"
        windows:
          type: array
          items:
            type: string
            pattern: "^[a-z][a-z0-9-]{0,31}$"
    AssistantMessageRequest:
      type: object
      required: [content, current_window, terminals]
      properties:
        content:
          type: string
        current_node:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,31}$"
        current_window:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,31}$"
        terminals:
          type: array
          items:
            $ref: "#/components/schemas/AssistantTerminalContext"
...